backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
openai==1.99.9
python-dotenv==1.1.1
beautifulsoup4==4.13.4
cryptography==44.0.3
